+++ Sources/Montage/Element/Button/TextButton/UITextButton.swift
@@ -337,68 +337,3 @@ extension Button.UITextButton: UIGestureRecognizerDelegate {
         return true
     }
 }
-
-import SwiftUI
-
-extension Button.TextButton.Variant {
-    var activeUIColor: UIColor {
-        switch self {
-        case .primary: .alias(.primaryNormal)
-        case .assistive: .alias(.labelAlternative).withAlphaComponent(0.61)
-        }
-    }
-    
-    var activeColor: SwiftUI.Color {
-        .init(uiColor: activeUIColor)
-    }
-    
-    var inactiveUIColor: UIColor {
-        .alias(.labelDisable).withAlphaComponent(0.16)
-    }
-    
-    var inactiveColor: SwiftUI.Color {
-        .init(uiColor: inactiveUIColor)
-    }
-
-    var interactionRadius: CGFloat {
-        6.0
-    }
-    
-    var interactionVariant: Decorate.Interaction.Variant {
-        switch self {
-        case .primary: .normal
-        case .assistive: .light
-        }
-    }
-    
-    var interactionColor: Color.Alias {
-        switch self {
-        case .primary: .primaryNormal
-        case .assistive: .labelNormal
-        }
-    }
-}
-
-extension Button.TextButton.Size {
-    var iconSize: CGSize {
-        switch self {
-        case .medium:
-            return .init(width: 20, height: 20)
-        case .small:
-            return .init(width: 16, height: 16)
-        }
-    }
-    
-    var typoVariant: Typography.Variant {
-        switch self {
-        case .medium:
-            return .body1
-        case .small:
-            return .label1
-        }
-    }
-    
-    var typoWeight: Typography.Weight {
-        return .bold
-    }
-}
